Bug fixes, player lives
Fixed a bug where I misspelled a file name.... Also, a bug on the Name
enter screen where holding space adds spaces, so I truncated the name.

Added in support for player lives. Depending on the number in Misc2 and
whether Switch0 is active or not, players can get more than one life!