Screen Shake
Finally got it implemented!!! That was oddly difficult :( Mainly, it was
the timing and getting the screen back to center that was stumping me.

It's working, and it's a nice visual effect! Makes it harder to play the
game >:P

I was thinking of adding it to when a player uses a bomb, but like...
hrm